requirements.txt
langchain_core
langchain_google_genai
flask
python-pptx
python-dotenv
gunicorn